Remove usage of codecov ATS feature
We are in the process of deprecating the ATS feature of codecov, and I saw you are one of the few (open source) users of it.

This PR removes the usage of ATS, just calling out to the whole testsuite all at once.
I did keep some of the `--collect` logic for the `codemod-tests` CI job, as I believe this is doing some kind of testsuite splitting.

As I saw you have modified the CI related workflows fairly recently, I’m curious whether the ATS feature as it exists is providing any value to you?